requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
requests
Pillow
python-dotenv
pytest
pytest-flask
gunicorn
psycopg2-binary
google-generativeai